requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-groq
langchain-community
duckduckgo-search
requests